pyproject.toml
[project]
name = "cyberbullying-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Hector", email = "data_analitics_HLH@protonmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "ipywidgets>=8.1.5",
    "jupyterlab>=4.3.4",
    "langchain>=0.3.12",
    "mypy>=1.14.0",
    "pandas>=2.2.3",
    "pandera[polars]>=0.22.1",
    "polars>=1.17.1",
    "pyarrow>=18.1.0",
    "pydantic>=2.10.3",
    "scalene>=1.5.49",
    "seaborn>=0.13.2",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.47.1",
]
[tool.hatch.build.targets.wheel]
 packages = ["src/cyberbullying-detection"]
[project.scripts]
cyberbullying-detection = "cyberbullying_detection:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
